uart: move under experimental/conn/
It is not ready for public consumption yet and that would be confusing users.
